Call location service in app component
